+++ App/src/components/ListRenderer.tsx
@@ -5,22 +5,22 @@ import { StyleSheet, View } from 'react-native';
 import { Divider, Theme } from 'react-native-paper';
 import { IMemberOverviewFragment } from "../model/IMemberOverviewFragment";
 import { Section } from './List/Section';
-import { MemberListItem } from "./Member/MemberListItem";
+import { MemberItemWithSwipe } from './Member/MemberItemWithSwipe';
 
 export function renderItem(
     item: IMemberOverviewFragment | string,
     theme: Theme,
     onPress?: (item: IMemberOverviewFragment) => void,
-    margin?: number,
+    margin: number = 30,
     height?: number,
-    ) {
+) {
     if (typeof (item) === "string") {
         return <Section title={item} backgroundColor={theme.colors.primary} />
     }
 
     return (
         <>
-            <MemberListItem height={height} theme={theme} onPress={onPress} member={item} margin={margin} />
+            <MemberItemWithSwipe height={height} theme={theme} onPress={onPress} item={item} margin={margin} />
             {renderDivider(theme)}
         </>
     );