+++ App/assets/docs/releasenotes.md
@@ -7,6 +7,7 @@
 - Nearby Members shows other members around you, that are willing to share their location. Get connected!
 - Readability of text and html has been optimized for dark theme, #54
 - Expandable sections on club page now only prompt for expansion with enough elements
+- Favorites now need swipe-right to toggle value
 
 # v1.0.1